refactor: CarsDto 구현 및 Winners를 DTO로 수정
- Cars의 데이터를 가져오는 CarsDto 클래스 구현.
- Winners의 결과 반환 로직을 서비스 내부에서 수행.
- Winners의 클래스명을 WinnersResult로 변경.
