+++ src/main/java/racingcar/controller/Controller.java
@@ -2,7 +2,7 @@ package racingcar.controller;
 
 import racingcar.dto.AttemptCount;
 import racingcar.dto.CarsState;
-import racingcar.dto.WinnersDto;
+import racingcar.dto.WinnersResult;
 import racingcar.model.Cars;
 import racingcar.service.GamePlayingService;
 import racingcar.view.View;
@@ -56,7 +56,7 @@ public class Controller {
         view.printFinalWinnerMessage(getWinners(cars));
     }
 
-    private WinnersDto getWinners(Cars cars) {
+    private WinnersResult getWinners(Cars cars) {
         return gamePlayingService.getWinners(cars);
     }
 }
