+++ src/main/java/racingcar/dto/CarsDto.java
@@ -1,0 +1,7 @@
+package racingcar.dto;
+
+import java.util.List;
+import racingcar.model.Car;
+
+public record CarsDto(List<Car> cars) {
+}
