--- src/main/java/racingcar/dto/WinnersDto.java
@@ -1,4 +1,0 @@
-package racingcar.dto;
-
-public record WinnersDto(String winners) {
-}
