+++ src/main/java/racingcar/dto/WinnersResult.java
@@ -1,0 +1,4 @@
+package racingcar.dto;
+
+public record WinnersResult(String winners) {
+}
