+++ src/main/java/racingcar/service/GamePlayingService.java
@@ -1,17 +1,16 @@
 package racingcar.service;
 
+import java.util.List;
+import racingcar.Constants;
 import racingcar.dto.AttemptCount;
+import racingcar.dto.CarsDto;
 import racingcar.dto.CarsState;
-import racingcar.dto.WinnersDto;
+import racingcar.dto.WinnersResult;
+import racingcar.model.Car;
 import racingcar.model.Cars;
 import racingcar.validator.AttemptCountValidator;
-import racingcar.vo.Winners;
 
 public class GamePlayingService {
-    public GamePlayingService() {
-
-    }
-
     public Cars generateCars(String inputCars) {
         return new Cars(inputCars);
     }
@@ -30,11 +29,33 @@ public class GamePlayingService {
         return cars.getCurrentCarsResult();
     }
 
-    public WinnersDto getWinners(Cars cars) {
-        return toWinnersDto(cars.findWinners());
+    public WinnersResult getWinners(Cars cars) {
+        CarsDto winners = cars.findWinners();
+        return toWinners(winners.cars());
     }
 
-    private WinnersDto toWinnersDto(Winners result) {
-        return result.toDto();
+    private WinnersResult toWinners(List<Car> winners) {
+        StringBuilder result = new StringBuilder();
+
+        winners.stream().map(Car::getNameValue)
+                .forEach((name) -> addResult(result, name));
+
+        trim(result);
+        return new WinnersResult(result.toString());
+    }
+
+    private void addResult(StringBuilder result, String name) {
+        result.append(Constants.WHITESPACE);
+        result.append(name);
+        result.append(Constants.COMMA);
+    }
+
+    private void trim(StringBuilder result) {
+        result.deleteCharAt(Constants.RESULT_FIRST_INDEX);
+        result.deleteCharAt(getLastIndexOfResult(result));
+    }
+
+    private int getLastIndexOfResult(StringBuilder result) {
+        return result.length() - 1;
     }
 }
